Move update tests to its own file
